no random barcode present possible
change behavior of trim_random_barcode:
If length of random barcode is set to 0, the entire read is kept. This
can be useful if reads are already demultiplexed but you want to use
the pipeline to process the samples. In that case, merge all samples
using cat and run pipeline